refactor: move communication dialogs to mixin
